requirements.txt
# HTS-CWT-AE-GW Requirements (synthetic-only, no LIGO stack)

# Core
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
PyYAML>=6.0

# CWT and preprocessing
PyWavelets>=1.1.0

# Evaluation and plotting
scikit-learn
matplotlib>=3.4.0
seaborn>=0.11.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
